requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-socketio[asgi]==5.11.3
pydantic==2.9.2
anyio==4.6.2.post1
